Refactoring to support horizontal view + new appfooter
some of this code is temporary as refactoring has to be made to the TitleBar to better support horizontal view. Additionally, the filter icons should anchor upon scrolling so that more real estate is made available to the device list. Finally, the color scheme should be revisited as the existing LightSteelBlue on the cards is extremely vulgar to the eyes.